Improvements to saving window size and position
On single window apps, window settings were not being saved because the app is destroyed by OF before the window, potentially causing a crash on exit because the closing window tries to access the MTApp instance on close. The solution was to not wait until exit to save app preferences, instead doing so whenever a window changed position or size. That required for us to listen to the GLFWWindow window moved event, because for some reason OF does not do so.